requirements.txt
Flask
Flask-SQLAlchemy
google-auth-oauthlib
requests
python-dotenv 